Move demo-warning file
