+++ layouts/editor.js
@@ -16,7 +16,7 @@ import WelcomeMessage from '@/components/welcome-message'
 import CertificationMessage from '@/components/certification-message'
 import Settings from '@/components/settings'
 import AddressEditor from '@/components/bal/address-editor'
-import DemoWarning from '@/components/sub-header/demo-warning'
+import DemoWarning from '@/components/demo-warning'
 
 function Editor({baseLocale, commune, voie, toponyme, voies, toponymes, numeros, children}) {
   const [isHidden, setIsHidden] = useState(false)
